feat: move add module searchbox to right panel
