+++ webui/src/Instances/index.jsx
@@ -3,10 +3,14 @@ import { useCallback, useContext, useRef } from "react";
 import { HelpModal } from "./HelpModal";
 import { CompanionContext, socketEmit } from "../util";
 import { InstancesList } from "./InstanceList";
+import { AddInstancesPanel } from "./AddModule";
+import { InstanceEditModal } from "./InstanceEditModal";
 
 export function InstancesPage({ resetToken }) {
 	const context = useContext(CompanionContext)
+
 	const helpModalRef = useRef()
+	const editModalRef = useRef()
 
 	const showHelp = useCallback((name) => {
 		socketEmit(context.socket, 'instance_get_help', [name]).then(([err, result]) => {
@@ -20,15 +24,21 @@ export function InstancesPage({ resetToken }) {
 		})
 	}, [context.socket])
 
+	const doConfigureInstance = useCallback((id) => {
+		editModalRef.current.show(id)
+	}, [])
+
 	return (
 		<CRow className='instances-page'>
 			<HelpModal ref={helpModalRef} />
 
+			<InstanceEditModal ref={editModalRef} />
+
 			<CCol xl={6} className='instances-panel'>
-				<InstancesList showHelp={showHelp} />
+				<InstancesList showHelp={showHelp} doConfigureInstance={doConfigureInstance} />
 			</CCol>
 			<CCol xl={6} className='instances-panel'>
-				<p> TODO - add bits will go here</p>
+				<AddInstancesPanel showHelp={showHelp} doConfigureInstance={doConfigureInstance} />
 			</CCol>
 		</CRow>
 	)
